refactor: split next.config.ts into config fragments
- Extract content-security-policy builder into config/content-security-policy.ts
- Extract redirects into config/redirects.ts with strict Redirect type
- Extract headers into config/headers.ts with strict Header type
- Extract SVG webpack/turbopack config into config/svg.ts
- next.config.ts now imports and composes all fragments
- Expand biome lint scope from src/ to entire project root
