+++ package.json
@@ -7,9 +7,9 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "biome check src/ && node scripts/check-server-only-imports.mjs",
-    "lint:fix": "biome check --write src/",
-    "format": "biome format --write src/",
+    "lint": "biome check . && node scripts/check-server-only-imports.mjs",
+    "lint:fix": "biome check --write .",
+    "format": "biome format --write .",
     "test": "vitest run",
     "type-check": "tsc --noEmit",
     "env:pull": "vercel env pull .env.local",
