+++ tailwind.config.ts
@@ -60,8 +60,8 @@ export default {
       white: "#ffffff",
     },
     fontFamily: {
-      sans: ['var(--font-geist-sans)', 'system-ui', 'sans-serif'],
-      mono: ['var(--font-geist-mono)', 'monospace'],
+      sans: ["var(--font-geist-sans)", "system-ui", "sans-serif"],
+      mono: ["var(--font-geist-mono)", "monospace"],
     },
     extend: {
       colors: {
